requirements.txt
pyTelegramBotAPI
python-telegram-bot
python-dotenv
redis
rq
httpx
flask
